refactor: Rename pdbml_enums module to schemas for better organization
Rename sloth/pdbml_enums.py to sloth/schemas.py to better organize the module
structure and align with common naming conventions for schema-related constants.

### 🔧 Core Changes:

**File Operations:**
- Renamed sloth/pdbml_enums.py → sloth/schemas.py (file rename, no content changes)
- Updated import statements in 3 files to use schemas module

**Updated Import Statements in:**
- sloth/pdbml_converter.py: XML converter enum imports
- tests/test_imports.py: Test enum imports and functionality tests
- tests/test_pipeline_integration.py: Enum class functionality tests

**Enum Classes Relocated:**
- XMLLocation: XML element location types
- XMLElementType: XML element structural types
- XMLGroupingType: XML grouping strategies
- XMLContainerType: XML container types
- PDBMLElement: Standard PDBML XML element names
- PDBMLAttribute: Standard PDBML XML attribute names
- DebugFile: Debug file name constants
- NumericDataType: mmCIF numeric data types
- MMCIFDataType: Complete mmCIF data type system
- PDBMLNamespace: XML namespace constants
- NullValue: Null value representations

**Helper Functions Relocated:**
- get_numeric_fields(): Schema-driven numeric field detection
- is_null_value(): Null value checking utility

### 📂 Files Modified: 4 files
- 1 file renamed: pdbml_enums.py → schemas.py
- 3 files updated: Import statement changes in converter and tests

### 🎯 Benefits:

1. **Better Organization**: Consolidates schema-related constants under intuitive naming
2. **Naming Consistency**: Aligns with common conventions (schemas for constants/definitions)
3. **Clearer Purpose**: Module name better reflects its role as schema/constant definitions
4. **Improved Navigation**: Easier to locate schema-related functionality

### ✅ Quality Assurance:

- Pure rename operation - no functional changes
- All imports updated consistently
- Test coverage maintained for all enum functionality
- Schema-driven approach preserved

This refactoring improves module organization by using conventional naming
for schema and constant definitions while maintaining all
