+++ sloth/pdbml_converter.py
@@ -20,7 +20,7 @@ from functools import lru_cache, wraps
 from .models import MMCIFDataContainer, DataBlock, Category
 from .parser import MMCIFParser
 from .validators import XMLSchemaValidator
-from .pdbml_enums import (
+from .schemas import (
     XMLLocation, XMLElementType, XMLGroupingType, XMLContainerType,
     PDBMLElement, PDBMLAttribute, DebugFile, get_numeric_fields, 
     is_null_value, PDBMLNamespace
